adjust cmake

+++ cmake/export.sh
@@ -12,15 +12,25 @@ LIB_NAME=${LIB_NAME:-"xxxx"}
 
 echo "LIB_NAME: $LIB_NAME"
 
+if [ -d $LIB_NAME ]; then
+	echo "$LIB_NAME directory already exists."
+	exit
+fi
+
 cp -arp ./temp-lib $LIB_NAME
 
 cd $LIB_NAME
 
 find -name "CMakeLists.txt" | xargs sed -i "s/TEMP/$LIB_NAME/g"
 
-for file in `find -name "*TEMP*"`
+for file in `find . -name "*TEMP*"`
 do
 	echo "--$file"
+	echo "$file" | grep "package"
+	if [ $? -eq 0 ]; then
+		continue
+	fi
+	echo "Edit ..."
 	rm $file
 	new_file=`echo "$file" | sed "s/TEMP/$LIB_NAME/g"`
 	vim $new_file
